requirements.txt
asgiref==3.4.1
Django==3.2.8
django-jalali-date==1.0.0
jdatetime==3.6.4
pytz==2021.3
sqlparse==0.4.2
